requirements.txt
mcp>=1.27.0
sounddevice>=0.5.0
numpy>=1.24.0
praat-parselmouth>=0.4.3
